pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ioc-enricher"
version = "0.1.0"
description = "IOC enrichment tool"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "requests>=2.31",
  "python-dotenv>=1.0",
]

[project.optional-dependencies]
test = [
  "pytest>=7.4",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.pytest.ini_options]
testpaths = ["tests"]